+++ package.json
@@ -24,6 +24,7 @@
 		"typescript": "4.7.4"
 	},
 	"dependencies": {
+		"json-2-csv": "^5.0.1",
 		"ts-md5": "^1.3.1"
 	}
 }
