+++ src/converter/ViewToContentConverter.ts
@@ -1,0 +1,8 @@
+import {CodeBlockPostViewInfo} from "../model/ViewMesModel";
+
+export interface ViewToContentConverter {
+
+	supportViewInputType(): string;
+
+	convert(view: CodeBlockPostViewInfo): Promise<string>;
+}
