+++ src/main.ts
@@ -58,16 +58,53 @@ export default class CodeBlockTemplatePlugin extends Plugin {
 
         // 会在第二个view加载时，el还没被渲染，第一个view所以会因为无法获取el而被删除。通过添加isInit判断，可以避免这个问题
         if(this.isInit && !this.viewManager.getIDList4Name(viewName).contains(viewId)) this.viewManager.deleteInvalidId(viewName);
-        
-        this.viewManager.updateView(viewName, viewId, {
-          input: source,
-          viewPath: ctx.sourcePath,
-        })
+
+		  const view = {
+			  id: viewId,
+			  name: viewName,
+			  input: source,
+			  inputType: "text",
+			  viewPath: ctx.sourcePath,
+		  };
+
+        this.viewManager.updateView(viewName, viewId, view)
         
 
-        if(this.isInit) this.templateRender.render4ID(viewName, viewId, source, ctx.sourcePath);
+        if(this.isInit) this.templateRender.render4ID(view);
       },
     )
+
+	  this.registerMarkdownCodeBlockProcessor(
+		  'pack-view-csv',
+		  (source, el, ctx) => {
+			  // __________________获取viewName__________________
+			  const viewName = this.sourceManager.getCodeBlockIdentifier4View(ctx,el)
+			  if (viewName === undefined) return;
+
+			  // __________________将TemplContent渲染到页面__________________
+			  // 每个view都加上class和no属性
+
+			  // 通过文件和行号创建唯一的viewID
+			  const viewId = this.viewManager.createID(this.app.workspace.getActiveFile(), ctx.getSectionInfo(el)?.lineStart.toString() ?? "error")
+
+			  el.addClass(viewName, `pack-view-${viewName}-${viewId}`);
+
+			  // 会在第二个view加载时，el还没被渲染，第一个view所以会因为无法获取el而被删除。通过添加isInit判断，可以避免这个问题
+			  if(this.isInit && !this.viewManager.getIDList4Name(viewName).contains(viewId)) this.viewManager.deleteInvalidId(viewName);
+
+			  const view = {
+				  id: viewId,
+				  name: viewName,
+				  input: source,
+				  inputType: "csv",
+				  viewPath: ctx.sourcePath,
+			  };
+			  this.viewManager.updateView(viewName, viewId, view)
+
+
+			  if(this.isInit) this.templateRender.render4ID(view);
+		  },
+	  )
 
     this.registerMarkdownCodeBlockProcessor(
       'pack-source',
